requirements.txt
anthropic>=0.40.0
tavily-python>=0.3.0
PyGithub>=2.1.1
requests>=2.31.0
python-dotenv>=1.0.0
